labs/4-IoT/web_app/requirements.txt
google-cloud-bigquery
db-dtypes
Flask
requests
